+++ packages/atomic/cypress/integration/result-list/result-components/result-link.cypress.ts
@@ -120,7 +120,7 @@ describe('Result Link Component', () => {
         setupResultLink({slots: [slottedComponent]});
       });
 
-      it('should render the slot inside of the "a" tag', () => {
+      it('should render the default slot inside of the "a" tag', () => {
         ResultLinkSelectors.firstInResult()
           .find('a #myslot')
           .should('be.visible');
@@ -129,14 +129,16 @@ describe('Result Link Component', () => {
 
     describe('when there is a valid slot named "attributes"', () => {
       it('copies the attributes properly', () => {
-        const slot = generateComponentHTML('a', {
-          download: '',
+        const attributesSlot = generateComponentHTML('a', {
+          download: 'test',
           target: '_self',
+          slot: 'attributes',
         });
-        setupResultLink({slots: [slot]});
+        setupResultLink({slots: [attributesSlot]});
         ResultLinkSelectors.firstInResult()
-          .find('a')
-          .should('have.attr', 'download', '')
+          .find('a:not([slot])')
+          .should('have.attr', 'download', 'test')
+          .and('be.visible')
           .and('have.attr', 'target', '_self');
       });
     });
@@ -144,8 +146,9 @@ describe('Result Link Component', () => {
     describe('when there is a slot named "attributes" & a default slot', () => {
       before(() => {
         const attributesSlot = generateComponentHTML('a', {
-          download: '',
+          download: 'test',
           target: '_self',
+          slot: 'attributes',
         });
         const defaultSlot = document.createElement('div');
         defaultSlot.id = 'myslot';
@@ -156,12 +159,13 @@ describe('Result Link Component', () => {
 
       it('copies the attributes properly', () => {
         ResultLinkSelectors.firstInResult()
-          .find('a')
-          .should('have.attr', 'download', '')
+          .find('a:not([slot])')
+          .should('have.attr', 'download', 'test')
+          .and('be.visible')
           .and('have.attr', 'target', '_self');
       });
 
-      it('should render the slot inside of the "a" tag', () => {
+      it('should render the default slot inside of the "a" tag', () => {
         ResultLinkSelectors.firstInResult()
           .find('a #myslot')
           .should('be.visible');
@@ -171,8 +175,9 @@ describe('Result Link Component', () => {
     describe('when there is a slot named "attributes" & a default slot (empty string)', () => {
       before(() => {
         const attributesSlot = generateComponentHTML('a', {
-          download: '',
+          download: 'test',
           target: '_self',
+          slot: 'attributes',
         });
         const defaultSlot = document.createElement('div');
         defaultSlot.id = 'myslot';
@@ -184,12 +189,13 @@ describe('Result Link Component', () => {
 
       it('copies the attributes properly', () => {
         ResultLinkSelectors.firstInResult()
-          .find('a')
-          .should('have.attr', 'download', '')
+          .find('a:not([slot])')
+          .should('have.attr', 'download', 'test')
+          .and('be.visible')
           .and('have.attr', 'target', '_self');
       });
 
-      it('should render the slot inside of the "a" tag', () => {
+      it('should render the default slot inside of the "a" tag', () => {
         ResultLinkSelectors.firstInResult()
           .find('a #myslot')
           .should('be.visible');