+++ packages/atomic/src/components/insight/atomic-insight-search-box/atomic-insight-search-box.tsx
@@ -1,9 +1,4 @@
 import {Component, h, Prop, State} from '@stencil/core';
-import {
-  buildSearchBox,
-  SearchBox,
-  SearchBoxState,
-} from '@coveo/headless/insight';
 import {randomID} from '../../../utils/utils';
 import {
   BindStateToController,
@@ -13,6 +8,11 @@ import {InsightBindings} from '../atomic-insight-interface/atomic-insight-interf
 import {SearchInput} from '../../common/search-box/search-input';
 import {SearchBoxWrapper} from '../../common/search-box/search-box-wrapper';
 import {SubmitButton} from '../../common/search-box/submit-button';
+import {
+  buildInsightSearchBox,
+  InsightSearchBox,
+  InsightSearchBoxState,
+} from '..';
 
 /**
  *
@@ -28,7 +28,7 @@ export class AtomicInsightSearchBox {
   @State() public error!: Error;
 
   private id!: string;
-  private searchBox!: SearchBox;
+  private searchBox!: InsightSearchBox;
   private inputRef!: HTMLInputElement;
 
   /**
@@ -39,7 +39,7 @@ export class AtomicInsightSearchBox {
 
   @BindStateToController('searchBox')
   @State()
-  private searchBoxState!: SearchBoxState;
+  private searchBoxState!: InsightSearchBoxState;
 
   public initialize() {
     this.id = randomID('atomic-search-box-');
@@ -58,7 +58,7 @@ export class AtomicInsightSearchBox {
         },
       },
     };
-    this.searchBox = buildSearchBox(this.bindings.engine, {
+    this.searchBox = buildInsightSearchBox(this.bindings.engine, {
       options: searchBoxOptions,
     });
   }