+++ packages/atomic/src/components/search/atomic-search-box/atomic-search-box.tsx
@@ -37,10 +37,11 @@ import {AriaLiveRegion} from '../../../utils/accessibility-utils';
 import {SafeStorage, StorageItems} from '../../../utils/local-storage-utils';
 import {promiseTimeout} from '../../../utils/promise-utils';
 import {updateBreakpoints} from '../../../utils/replace-breakpoint';
+import {Bindings} from '../atomic-search-interface/atomic-search-interface';
 import {SearchInput} from '../../common/search-box/search-input';
 import {SearchBoxWrapper} from '../../common/search-box/search-box-wrapper';
 import {SubmitButton} from '../../common/search-box/submit-button';
-import {Bindings} from '../atomic-search-interface/atomic-search-interface';
+
 /**
  * The `atomic-search-box` component creates a search box with built-in support for suggestions.
  *
@@ -372,7 +373,7 @@ export class AtomicSearchBox {
       if (prom.status === 'fulfilled') {
         fulfilledSuggestions.push(this.suggestions[j]);
       } else {
-        console.error(
+        this.bindings.engine.logger.warn(
           'Some query suggestions are not being shown because the promise timed out.'
         );
       }