+++ packages/samples/atomic-react/src/pages/InstantResultsPage.tsx
@@ -1,5 +1,4 @@
 import React, {FunctionComponent} from 'react';
-import {AtomicResultLink, AtomicResultList} from '@coveo/atomic-react';
 import {AtomicPageWrapper} from '../components/AtomicPageWrapper';
 
 export const InstantResultsPage: FunctionComponent = () => {
@@ -9,12 +8,7 @@ export const InstantResultsPage: FunctionComponent = () => {
       organizationId="electronicscoveodemocomo0n2fu8v"
       options={{recentQueries: true, instantResults: true}}
     >
-      <AtomicResultList
-        fieldsToInclude="ec_price,ec_rating,ec_images,ec_brand,cat_platform,cat_condition,cat_categories,cat_review_count,cat_color"
-        display="grid"
-        imageSize="large"
-        template={() => <AtomicResultLink />}
-      />
+      <p>No result list</p>
     </AtomicPageWrapper>
   );
 };