+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,87 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.94.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.93.2...@coveo/atomic@1.94.0) (2022-08-02)
+
+
+### Bug Fixes
+
+* **atomic:** change timeout message from error to warning ([#2276](https://github.com/coveo/ui-kit/issues/2276)) ([4a8eecc](https://github.com/coveo/ui-kit/commit/4a8eeccd01dcf76d085277a7d74169c3d99dbdea))
+
+
+### Features
+
+* **atomic:** add insight refine toggle and insight refine modal ([#2268](https://github.com/coveo/ui-kit/issues/2268)) ([d742a51](https://github.com/coveo/ui-kit/commit/d742a517dcb29d42a9d898a22a448b912f628811))
+* **atomic:** scrollable tests ([#2225](https://github.com/coveo/ui-kit/issues/2225)) ([fca12a7](https://github.com/coveo/ui-kit/commit/fca12a7a81582011136546667235b8ad5fc7f1a3))
+
+
+
+
+
+## [1.93.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.93.1...@coveo/atomic@1.93.2) (2022-08-01)
+
+
+### Bug Fixes
+
+* **atomic:** ensured aria labels contain innerText ([#2255](https://github.com/coveo/ui-kit/issues/2255)) ([2f641c7](https://github.com/coveo/ui-kit/commit/2f641c70ad780d696c0fcfff90c4266ebdef8a9a))
+
+
+
+
+
+## [1.93.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.93.0...@coveo/atomic@1.93.1) (2022-08-01)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.93.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.92.0...@coveo/atomic@1.93.0) (2022-07-29)
+
+
+### Features
+
+* **atomic:** fix scrollable mobile view overscrolling ([#2264](https://github.com/coveo/ui-kit/issues/2264)) ([d5632a4](https://github.com/coveo/ui-kit/commit/d5632a4bcea73fde715971a82c739eca5612ece6))
+
+
+
+
+
+# [1.92.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.91.0...@coveo/atomic@1.92.0) (2022-07-29)
+
+
+### Features
+
+* **atomic:** add edit and history buttons to insight ([#2258](https://github.com/coveo/ui-kit/issues/2258)) ([bc0b928](https://github.com/coveo/ui-kit/commit/bc0b92865b3c25598de50cfc22eccb52a75568a3))
+
+
+
+
+
+# [1.91.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.90.2...@coveo/atomic@1.91.0) (2022-07-28)
+
+
+### Features
+
+* **atomic,headless:** add result list and result template to insight ([#2197](https://github.com/coveo/ui-kit/issues/2197)) ([767b242](https://github.com/coveo/ui-kit/commit/767b24200e8fc43b64fb85075a97b574de5bc9ba))
+
+
+
+
+
+## [1.90.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.90.1...@coveo/atomic@1.90.2) (2022-07-28)
+
+
+### Bug Fixes
+
+* **atomic:** pass rendering function down to instant result atomic-re… ([#2257](https://github.com/coveo/ui-kit/issues/2257)) ([03336d6](https://github.com/coveo/ui-kit/commit/03336d6d2fc07561aa8dc88e2e2b6c99228d9315))
+* **atomic:** result link default slot should not create a warning ([#2246](https://github.com/coveo/ui-kit/issues/2246)) ([3a7c133](https://github.com/coveo/ui-kit/commit/3a7c133f856e11651f58dc82eb1301a3681d425a))
+
+
+
+
+
 ## [1.90.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.90.0...@coveo/atomic@1.90.1) (2022-07-27)
 
 **Note:** Version bump only for package @coveo/atomic